// src/redux/draftConfirm/types/action.ts
// src/redux/draftConfirm/types/action.ts

// Action Types
export enum DraftConfirmActionTypes {
    // Fetch Orders
    FETCH_ORDERS_REQUEST = '@DRAFT_CONFIRM/FETCH_ORDERS_REQUEST',
    FETCH_ORDERS_SUCCESS = '@DRAFT_CONFIRM/FETCH_ORDERS_SUCCESS',
    FETCH_ORDERS_FAILURE = '@DRAFT_CONFIRM/FETCH_ORDERS_FAILURE',
    
    // Fetch Order Details
    FETCH_ORDER_DETAILS_REQUEST = '@DRAFT_CONFIRM/FETCH_ORDER_DETAILS_REQUEST',
    FETCH_ORDER_DETAILS_SUCCESS = '@DRAFT_CONFIRM/FETCH_ORDER_DETAILS_SUCCESS',
    FETCH_ORDER_DETAILS_FAILURE = '@DRAFT_CONFIRM/FETCH_ORDER_DETAILS_FAILURE',
    
    // Fetch CodeR List
    FETCH_CODE_R_REQUEST = '@DRAFT_CONFIRM/FETCH_CODE_R_REQUEST',
    FETCH_CODE_R_SUCCESS = '@DRAFT_CONFIRM/FETCH_CODE_R_SUCCESS',
    FETCH_CODE_R_FAILURE = '@DRAFT_CONFIRM/FETCH_CODE_R_FAILURE',
    
    // Add Item
    ADD_ITEM_REQUEST = '@DRAFT_CONFIRM/ADD_ITEM_REQUEST',
    ADD_ITEM_SUCCESS = '@DRAFT_CONFIRM/ADD_ITEM_SUCCESS',
    ADD_ITEM_FAILURE = '@DRAFT_CONFIRM/ADD_ITEM_FAILURE',
    
    // Remove Item
    REMOVE_ITEM_REQUEST = '@DRAFT_CONFIRM/REMOVE_ITEM_REQUEST',
    REMOVE_ITEM_SUCCESS = '@DRAFT_CONFIRM/REMOVE_ITEM_SUCCESS',
    REMOVE_ITEM_FAILURE = '@DRAFT_CONFIRM/REMOVE_ITEM_FAILURE',
    
    // Confirm Draft Order
    CONFIRM_DRAFT_ORDER_REQUEST = '@DRAFT_CONFIRM/CONFIRM_DRAFT_ORDER_REQUEST',
    CONFIRM_DRAFT_ORDER_SUCCESS = '@DRAFT_CONFIRM/CONFIRM_DRAFT_ORDER_SUCCESS',
    CONFIRM_DRAFT_ORDER_FAILURE = '@DRAFT_CONFIRM/CONFIRM_DRAFT_ORDER_FAILURE',
    
    // Cancel Order
    CANCEL_ORDER_REQUEST = '@DRAFT_CONFIRM/CANCEL_ORDER_REQUEST',
    CANCEL_ORDER_SUCCESS = '@DRAFT_CONFIRM/CANCEL_ORDER_SUCCESS',
    CANCEL_ORDER_FAILURE = '@DRAFT_CONFIRM/CANCEL_ORDER_FAILURE',
    
    // Clear Selected Order
    CLEAR_SELECTED_ORDER = '@DRAFT_CONFIRM/CLEAR_SELECTED_ORDER'
  }